Disable BeginFrame in effects
In effect.cc, a manual operation would take over during effects processing. However, since we have no glOrtho there, we can't reliably ensure it renders correctly to screen. This caused rendering that was too large for the screen and once the effect was complete, the image would "snap" back into place. Just disable it for now.